requirements.txt
google-generativeai>=0.3.0
Pillow>=10.0.0
termcolor>=2.3.0
python-pptx>=0.6.21
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
